+++ .swiftpm/xcode/xcshareddata/xcschemes/WalletConnect.xcscheme
@@ -76,6 +76,20 @@
                ReferencedContainer = "container:">
             </BuildableReference>
          </BuildActionEntry>
+         <BuildActionEntry
+            buildForTesting = "YES"
+            buildForRunning = "YES"
+            buildForProfiling = "YES"
+            buildForArchiving = "YES"
+            buildForAnalyzing = "YES">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "KMS"
+               BuildableName = "KMS"
+               BlueprintName = "KMS"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </BuildActionEntry>
       </BuildActionEntries>
    </BuildAction>
    <TestAction
@@ -121,6 +135,16 @@
                BlueprintIdentifier = "RelayerTests"
                BuildableName = "RelayerTests"
                BlueprintName = "RelayerTests"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "KMSTests"
+               BuildableName = "KMSTests"
+               BlueprintName = "KMSTests"
                ReferencedContainer = "container:">
             </BuildableReference>
          </TestableReference>
