+++ Sources/KMS/Extensions/String.swift
@@ -2,11 +2,6 @@
 
 import Foundation
 
-extension String {
-    func toHexEncodedString(uppercase: Bool = true, prefix: String = "", separator: String = "") -> String {
-        return unicodeScalars.map { prefix + .init($0.value, radix: 16, uppercase: uppercase) } .joined(separator: separator)
-    }
-}
 
 extension String: GenericPasswordConvertible {
