+++ Sources/WalletConnectUtils/String.swift
@@ -1,0 +1,9 @@
+
+
+import Foundation
+
+public extension String {
+    func toHexEncodedString(uppercase: Bool = true, prefix: String = "", separator: String = "") -> String {
+        return unicodeScalars.map { prefix + .init($0.value, radix: 16, uppercase: uppercase) } .joined(separator: separator)
+    }
+}
